// src/model/Character.ts
export interface Character {
  id: string;
  nombre: string;
  altura: string;
  masa: string;
  color_pelo: string;
  color_piel: string;
  color_ojo: string;
  fecha_nacimiento: string;
  genero: string;
  planeta_hogar: string;
  pelicula: string[];
  especie: string[];
  vehiculo: string[];
  naves: string[];
}