// src/model/Specie.ts
export interface Specie {
  id: string;
  nombre: string;
  clasificacion: string;
  designacion: string;
  altura_promedio: string;
  colores_piel: string[];
  colores_cabello: string[];
  colores_ojo: string[];
  promedio_vida: string;
  planeta_hogar: string;
  idioma: string;
  personajes: string[];
  peliculas: string[];
}