// src/model/Vehicles.ts
export interface Vehicles {
  id: string;
  nombre: string;
  modelo: string;
  fabricante: string;
  costo_en_creditos: string;
  longitud: string;
  velocidad_maxima_atmosfera: string;
  tripulacion: string;
  pasajeros: string;
  capacidad_carga: string;
  consumibles: string;
  clase_vehiculo: string;
  peliculas: string[];
  pilotos: string[];
}