// src/model/Starship.ts
export interface Starship {
  id: string;
  nombre: string;
  modelo: string;
  fabricante: string;
  costo_en_creditos: string;
  longitud: string;
  velocidad_maxima_atmosfera: string;
  tripulacion: string;
  pasajeros: string;
  capacidad_carga: string;
  consumibles: string;
  clase_vehiculo: string;
  MGLT: string;
  calificacion_hiperimpulsor: string;
  peliculas: string[];
  pilotos: string[];
}